Move platform to config package
