+++ sconfig/native/src/main/scala/org/ekrich/config/impl/PlatformConfigFactory.scala
@@ -1,4 +1,4 @@
-package org.ekrich.config.impl
+package org.ekrich.config
 
 /**
   * [[ConfigFactory]] methods for Scala Native platform
